+++ src/models/emojis/validators.rs
@@ -5,7 +5,8 @@ use crate::errors::ValidationError;
 const EMOJI_NAME_RE: &str = r"^[\w.]+$";
 pub const EMOJI_MAX_SIZE: usize = 250 * 1000; // 250 kB
 pub const EMOJI_LOCAL_MAX_SIZE: usize = 50 * 1000; // 50 kB
-pub const EMOJI_MEDIA_TYPES: [&str; 2] = [
+pub const EMOJI_MEDIA_TYPES: [&str; 3] = [
+    "image/apng",
     "image/gif",
     "image/png",
 ];